refactor(editor): Detangle users store from setting store
